test(e2e): split e2e commands/setup
